Move interface methods into their mapper files
