+++ back/pkg/entityservice/location_mapper.go
@@ -1,7 +1,10 @@
 package entityservice
 
 import (
+	"strconv"
 	"studi-guide/ent"
+	"studi-guide/ent/location"
+	"studi-guide/ent/tag"
 	"studi-guide/pkg/navigation"
 )
 
@@ -50,5 +53,72 @@ func (r *EntityService) mapSectionArray(sections []Section) ([]*ent.Section, err
 
 	return entSections, nil
 
+}
+
+func (r *EntityService) GetAllLocations() ([]Location, error) {
+	entLoactions, err := r.client.Location.Query().
+		WithTags().
+		WithPathnode().
+		All(r.context)
+	if err != nil {
+		return nil, err
+	}
+
+	return r.locationArrayMapper(entLoactions), nil
+}
+
+func (r *EntityService) GetLocation(name, building, campus string) (Location, error) {
+
+	q := r.client.Location.Query().WithPathnode().WithTags().Where(location.NameEQ(name))
+
+	if len(building) > 0 {
+		// TODO implement building
+	}
+
+	if len(campus) > 0 {
+		// TODO implement campus
+	}
+
+	entLocation, err := q.First(r.context)
+	if err != nil {
+		return Location{}, err
+	}
+	return *r.locationMapper(entLocation), nil
+}
+
+func (r *EntityService) FilterLocations(name, tagStr, floor, building, campus string) ([]Location, error) {
+
+	query := r.client.Location.Query().
+		WithPathnode().WithTags()
+
+	if len(name) > 0 {
+		query = query.Where(location.NameContains(name))
+	}
+
+	if len(tagStr) > 0 {
+		query = query.Where(location.HasTagsWith(tag.NameContains(tagStr)))
+	}
+
+	if len(floor) > 0 {
+		iFloor, err := strconv.Atoi(floor)
+		if err != nil {
+			return nil, err
+		}
+		query = query.Where(location.FloorEQ(iFloor))
+	}
+
+	if len(building) > 0 {
+		// Todo query building
+	}
+
+	if len(campus) > 0 {
+		// Todo query campus
+	}
+
+	entLocations, err := query.All(r.context)
+	if err != nil {
+		return nil, err
+	}
+	return r.locationArrayMapper(entLocations), nil
 }
